dream-coast-60132/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
dj-database-url==0.5.0
Django==1.11.10
django-common-helpers==0.9.1
django-cors-headers==2.2.0
django-cron==0.5.1
djangorestframework==3.7.7
enum34==1.1.6
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
joblib==0.12.0
numpy==1.14.2
ordereddict==1.1
pathlib==1.0.1
Pillow==5.0.0
pipenv==11.8.0
psycopg2-binary==2.7.4
pycparser==2.18
pyOpenSSL==17.5.0
pytz==2018.3
requests==2.18.4
scikit-surprise==1.0.5
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
